requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
biip==4.0.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
defusedxml==0.7.1
dj-database-url==2.3.0
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==0.61.1
django-cors-headers==4.6.0
django-countries==7.6.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
frozenlist==1.4.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
multidict==6.1.0
oauthlib==3.2.2
packaging==24.1
phonenumbers==8.13.45
pillow==10.4.0
pluggy==1.5.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.3
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
sqlparse==0.5.1
stripe==11.2.0
twilio==9.3.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
whitenoise==6.8.2
yarl==1.12.1
